Code Revisions register_new_user_api.php
Title => RegisterUser.php: This was really my fault. I should have named these better.

Line 29/36: If a user did not set all the fields for this call, this is a bad request. INTERNAL_ERROR should be reserved for "we as the programmer did something wrong, sorry." BAD_REQUEST is supposed to represent "something you ask for is not correct."

Line 40: Queries should be explicit about which columns they are inserting to, so that if the schema ever changes, we don't have to go back and find/edit every single query for this table.